Fix the email delivery statistics that were not displaying the correct ammount of messages.
